+++ src/useValidation/ValidationAbortedError.ts
@@ -1,0 +1,1 @@
+export class ValidationAbortedError extends Error {}
